Splitting Ultima.Network.Shared.cs. 1 file per class.
