+++ dev/Ultima/Data/ContentItem.cs
@@ -1,40 +1,16 @@
 ﻿/***************************************************************************
- *   Shared.cs
- *
+ *   ContentItem.cs
+ *   Copyright (c) 2015 UltimaXNA Development Team
+ *   
  *   This program is free software; you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
  *   the Free Software Foundation; either version 3 of the License, or
  *   (at your option) any later version.
  *
  ***************************************************************************/
-using UltimaXNA.Core.Network;
 
-namespace UltimaXNA.Ultima.Network
+namespace UltimaXNA.Ultima.Data
 {
-    public enum Sex
-    {
-        Male = 0,
-        Female = 1
-    }
-
-    public enum Race
-    {
-        Human = 1,
-        Elf = 2
-    }
-
-    public class HouseRevisionState
-    {
-        public Serial Serial;
-        public int Hash;
-
-        public HouseRevisionState(Serial serial, int revisionHash)
-        {
-            Serial = serial;
-            Hash = revisionHash;
-        }
-    }
-
     public class ContentItem
     {
         public readonly Serial Serial;
@@ -56,20 +32,6 @@ namespace UltimaXNA.Ultima.Network
             GridLocation = gridLocation;
             ContainerSerial = containerSerial;
             Hue = hue;
-        }
-    }
-
-    public class StatLocks
-    {
-        public int Strength;
-        public int Dexterity;
-        public int Intelligence;
-
-        public StatLocks(int s, int d, int i)
-        {
-            Strength = s;
-            Dexterity = d;
-            Intelligence = i;
         }
     }
 }
