+++ dev/Ultima/Data/Genders.cs
@@ -1,0 +1,19 @@
+﻿/***************************************************************************
+ *   Genders.cs
+ *   Copyright (c) 2015 UltimaXNA Development Team
+ *   
+ *   This program is free software; you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation; either version 3 of the License, or
+ *   (at your option) any later version.
+ *
+ ***************************************************************************/
+
+namespace UltimaXNA.Ultima.Data
+{
+    public enum Genders
+    {
+        Male = 0,
+        Female = 1
+    }
+}
