+++ dev/Ultima/Network/Server/ContainerContentPacket.cs
@@ -1,13 +1,7 @@
 ﻿/***************************************************************************
  *   ContainerContentPacket.cs
+ *   Copyright (c) 2015 UltimaXNA Development Team
  *   
- *   begin                : May 31, 2009
- *   email                : poplicola@ultimaxna.com
- *
- ***************************************************************************/
-
-/***************************************************************************
- *
  *   This program is free software; you can redistribute it and/or modify
  *   it under the terms of the GNU General Public License as published by
  *   the Free Software Foundation; either version 3 of the License, or
@@ -18,6 +12,7 @@
 using System.Collections.Generic;
 using UltimaXNA.Core.Network;
 using UltimaXNA.Core.Network.Packets;
+using UltimaXNA.Ultima.Data;
 #endregion
 
 namespace UltimaXNA.Ultima.Network.Server
